requirements.txt
load_dotenv
Flask
slack_sdk
openai